Not discarding forks with failing promise check
When removing the extra self-star in stargazersCheckPromises, we
shouldn't add a repository to bad_forks if the network request fails. I
believe it is better to keep a repo in the list with an extra star than
throw out the whole repo just because our request failed.

Is there something I'm missing here?