Logic Flushed Out
This commit adds some commends and files that flesh out logic

We will use a sort of message queue where data queries
trigger immediate dom updates. This way new queries are
automatically loaded and processed.

We also will edit the dom to make it easier to supply changes to the
tree structure, and autmatic formatting with css selectors